# docs/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "CommentCountQuestion": {
        "additionalProperties": false,
        "properties": {
          "event_id": {
            "format": "uuid",
            "title": "Event Id",
            "type": "string"
          },
          "kind": {
            "const": "comment_count",
            "title": "Kind",
            "type": "string"
          },
          "question": {
            "maxLength": 300,
            "minLength": 2,
            "title": "Question",
            "type": "string"
          },
          "since": {
            "format": "date-time",
            "title": "Since",
            "type": "string"
          },
          "until": {
            "format": "date-time",
            "title": "Until",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "question",
          "event_id",
          "since",
          "until"
        ],
        "title": "CommentCountQuestion",
        "type": "object"
      },
...
      "ControlledCommentStatistics": {
        "properties": {
          "by_kind": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Kind",
            "type": "object"
          },
          "by_source": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Source",
            "type": "object"
          },
          "event_id": {
            "format": "uuid",
            "title": "Event Id",
            "type": "string"
          },
          "event_revision": {
            "minimum": 1.0,
            "title": "Event Revision",
            "type": "integer"
          },
          "rule_version": {
            "const": "event-comment-count-v1",
            "default": "event-comment-count-v1",
            "title": "Rule Version",
            "type": "string"
          },
          "since": {
            "format": "date-time",
            "title": "Since",
            "type": "string"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          },
          "until": {
            "format": "date-time",
            "title": "Until",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "event_revision",
          "since",
          "until",
          "total",
          "by_source",
          "by_kind"
        ],
        "title": "ControlledCommentStatistics",
        "type": "object"
      },
...
      "ErrorView": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "request_id"
        ],
        "title": "ErrorView",
        "type": "object"
      },
...
      "EvidenceQuestion": {
        "additionalProperties": false,
        "properties": {
          "event_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "kind": {
            "const": "evidence",
            "title": "Kind",
            "type": "string"
          },
          "limit": {
            "default": 3,
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "mode": {
            "default": "exact_substring",
            "enum": [
              "exact_substring",
              "semantic"
            ],
            "title": "Mode",
            "type": "string"
          },
          "question": {
            "maxLength": 300,
            "minLength": 2,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "question"
        ],
        "title": "EvidenceQuestion",
        "type": "object"
      },
...
      "KnowledgeAnswer": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeCitationView"
            },
            "title": "Citations",
            "type": "array"
          },
          "kind": {
            "enum": [
              "comment_count",
              "evidence"
            ],
            "title": "Kind",
            "type": "string"
          },
          "knowledge_entry_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Knowledge Entry Ids",
            "type": "array"
          },
          "method": {
            "enum": [
              "controlled_statistics_v1",
              "deterministic_retrieval_v1"
            ],
            "title": "Method",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "statistics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ControlledCommentStatistics"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "enum": [
              "answered",
              "unknown"
            ],
            "title": "Status",
            "type": "string"
          },
          "unknown_reason": {
            "anyOf": [
              {
                "const": "no_supported_evidence",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unknown Reason"
          }
        },
        "required": [
          "kind",
          "status",
          "question",
          "answer",
          "method"
        ],
        "title": "KnowledgeAnswer",
        "type": "object"
      },
      "KnowledgeCitationView": {
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "canonical_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Url"
          },
          "content_version_id": {
            "format": "uuid",
            "title": "Content Version Id",
            "type": "string"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "text_sha256": {
            "title": "Text Sha256",
            "type": "string"
          }
        },
        "required": [
          "content_version_id",
          "text_sha256",
          "text",
          "canonical_url",
          "available"
        ],
        "title": "KnowledgeCitationView",
        "type": "object"
      },
      "KnowledgeEntryView": {
        "properties": {
          "analysis_manifest_sha256": {
            "title": "Analysis Manifest Sha256",
            "type": "string"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeCitationView"
            },
            "minItems": 1,
            "title": "Citations",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "entry_type": {
            "const": "analysis_snapshot",
            "title": "Entry Type",
            "type": "string"
          },
          "event_id": {
            "format": "uuid",
            "title": "Event Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "semantic_index_state": {
            "enum": [
              "pending",
              "ready",
              "failed",
              "stale",
              "deleted"
            ],
            "title": "Semantic Index State",
            "type": "string"
          },
          "similarity": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": -1.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Similarity"
          },
          "source_analysis_run_id": {
            "format": "uuid",
            "title": "Source Analysis Run Id",
            "type": "string"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "event_id",
          "source_analysis_run_id",
          "entry_type",
          "version",
          "title",
          "body",
          "analysis_manifest_sha256",
          "semantic_index_state",
          "stale",
          "citations",
          "created_at",
          "updated_at"
        ],
        "title": "KnowledgeEntryView",
        "type": "object"
      },
      "KnowledgePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeEntryView"
            },
            "title": "Items",
            "type": "array"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "query_mode": {
            "default": "exact_substring",
            "enum": [
              "exact_substring",
              "semantic"
            ],
            "title": "Query Mode",
            "type": "string"
          }
        },
        "required": [
          "query",
          "items"
        ],
        "title": "KnowledgePage",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OwnerSession": {
        "in": "cookie",
        "name": "hk_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Single-owner learning workspace. Unsafe requests require an exact allowed Origin; authenticated writes also require a session-bound X-CSRF-Token. Monitor configuration history and zero-network query previews are available; collection remains disabled until a source passes rights and pipeline admission.",
    "title": "HotKey API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/analysis-runs/{identity}/knowledge-entry": {
      "post": {
        "operationId": "publishAnalysisKnowledge",
        "parameters": [
          {
            "in": "path",
            "name": "identity",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Identity",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeEntryView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "State conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Request content too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "OwnerSession": []
          }
        ],
        "summary": "Publish Analysis Knowledge",
        "tags": [
          "knowledge"
        ]
      }
    },
...
    "/api/knowledge": {
      "get": {
        "operationId": "searchKnowledge",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "minLength": 2,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "in": "query",
            "name": "event_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Id"
            }
          },
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "exact_substring",
              "enum": [
                "exact_substring",
                "semantic"
              ],
              "title": "Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Authentication required"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Request content too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "OwnerSession": []
          }
        ],
        "summary": "Search Knowledge",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/knowledge/query": {
      "post": {
        "operationId": "queryKnowledge",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "comment_count": "#/components/schemas/CommentCountQuestion",
                    "evidence": "#/components/schemas/EvidenceQuestion"
                  },
                  "propertyName": "kind"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CommentCountQuestion"
                  },
                  {
                    "$ref": "#/components/schemas/EvidenceQuestion"
                  }
                ],
                "title": "Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeAnswer"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Resource not found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Request content too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "OwnerSession": []
          }
        ],
        "summary": "Query Knowledge",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/knowledge/{identity}": {
      "get": {
        "operationId": "getKnowledgeEntry",
        "parameters": [
          {
            "in": "path",
            "name": "identity",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Identity",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeEntryView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Resource not found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Request content too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "OwnerSession": []
          }
        ],
        "summary": "Get Knowledge Entry",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/knowledge/{identity}/semantic-index": {
      "post": {
        "operationId": "indexKnowledgeEntry",
        "parameters": [
          {
            "in": "path",
            "name": "identity",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Identity",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeEntryView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "State conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Request content too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "OwnerSession": []
          }
        ],
        "summary": "Index Knowledge Entry",
        "tags": [
          "knowledge"
        ]
      }
    },
...
  }
}